feature(connectors):  modularized into core, factory and implementations
